pr fix

+++ backend/src/main/java/fr/projet/diginamic/backend/repositories/interfaces/ExpenseTypeRepository.java
@@ -7,6 +7,6 @@ import fr.projet.diginamic.backend.entities.ExpenseType;
 
 /** Repository of all ExpenseType method to search from database*/
 @Repository
-public interface ExpenseTypeRepository  extends JpaRepository<ExpenseType, Integer> {
+public interface ExpenseTypeRepository  extends JpaRepository<ExpenseType, Long> {
 
 }
